requirements.txt
astroid==1.6.3
attrs==17.4.0
autopep8==1.3.5
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
Django==2.0
django-cors-middleware==1.3.1
django-dotenv==1.4.2
django-oauth-toolkit==1.0.0
djangorestframework==3.8.1
Faker==0.8.8
future==0.16.0
idna==2.6
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mixer==6.0.1
more-itertools==4.1.0
mysqlclient==1.3.12
oauthlib==2.0.7
openapi-codec==1.3.2
pluggy==0.6.0
py==1.5.3
pycodestyle==2.3.1
psycopg2-binary==2.7.5
Pygments==2.2.0
pylint==1.8.4
pytest==3.5.0
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-env==0.6.2
python-dateutil==2.7.2
pytz==2018.3
requests==2.18.4
ruamel.yaml==0.15.37
six==1.11.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.22
wrapt==1.10.11
gunicorn==19.9.0
dj-database-url==0.5.0
djangorestframework-jwt==1.11.0
whitenoise==4.0
pyfcm==1.4.5
PyYAML==3.13
cloudinary==1.14.0